Правки 30.06

+++ app/Http/Livewire/Products.php
@@ -3,6 +3,7 @@
 namespace App\Http\Livewire;
 
 use App\Models\Product;
+use Illuminate\View\View;
 use Livewire\Attributes\On;
 use Livewire\WithPagination;
 use Livewire\Component;
@@ -11,15 +12,19 @@ class Products extends Component
 {
     use WithPagination;
 
-    public function delete(Product $product)
+    public function delete(Product $product): void
     {
-        $product->delete();
+        if ($product->orders()->count() === 0) {
+            $product->delete();
 
-        $this->dispatch('toast', message: __('Successfully deleted'));
+            $this->dispatch('toast', message: __('Successfully deleted'));
+        } else {
+            $this->dispatch('toast', message: __('You cannot delete an item because it has already been added to the order'), type: 'error');
+        }
     }
 
     #[On('echo:products,ProductEvent')]
-    public function render()
+    public function render(): View
     {
         $products = Product::orderByDesc('created_at')->paginate();
